MUS-550 : small fix (#162)
* MUS-550 : added  checkHipsterEligibility lambda

* MUS-550 : changed lambda name

* MUS-550 : changed name

* Update Jenkinsfile

* MUS-550 : corrected err

* MUS-550 : added test cases

* MUS-550 : changed logic

* MUS-550 : fixed messageCode

* Update Jenkinsfile

* MUS-550: corrected int

+++ lambdas/checkhipstereligibility/main.go
@@ -42,7 +42,7 @@ func handler(ctx context.Context, input pdwOutput) error {
 	status := "success"
 
 	if input.Status == "success" && len(input.Response.Data.Parcels) > 0 && input.Response.Data.Parcels[0].DetectedBuildingCount.Value != nil {
-		if buildingCount := input.Response.Data.Parcels[0].DetectedBuildingCount.Value.(int); buildingCount == 1 {
+		if buildingCount := input.Response.Data.Parcels[0].DetectedBuildingCount.Value.(float64); int(buildingCount) == 1 {
 			isHipsterCompatible = true
 		}
 	} else if input.Status == "failure" {
